refactor(groups): move claude.ai/design browse template into frontend/src/components/realty/browse
Move the 14 leaf components, the two page templates, and the format
helpers from docs/realty-groups/ to frontend/src/components/realty/browse/
so they become importable React modules. Extend frontend/src/types/realty.ts
with the directory wire types (RealtyGroupCard, GroupCardLayout,
GroupSidebarPlacement, GroupsSortKey, SortDirection, GroupMember, GroupReview)
matching the spec section 6.1 contract; hasVerifiedSlGroup is intentionally
omitted because the browse endpoint filters unverified groups server-side.
Drop the redundant local cn helper in favor of the project-wide @/lib/cn,
and extract formatFounded / initialsOf into the new
frontend/src/components/realty/browse/lib/format.ts so the shared cn utility
stays narrow.

Imports across the moved files are rewritten:
  ../types        -> @/types/realty
  ../lib/cn (cn)  -> @/lib/cn
  ../lib/cn (fmt) -> ../lib/format
  ../components/X -> ./components/X (from the page templates)

Task 9 in the same epic forks GroupsPage to drop client-side filtering and
externalize search/sort/page to props; Part 3 will fork GroupDetailPage for
the slug profile page. Those forks resolve the expected type errors in the
moved-but-unforked files (hasVerifiedSlGroup access, stale GroupsSortKey
enum values). No tests run in this task per the task contract.
