--- docs/realty-groups/index.ts
@@ -1,39 +1,0 @@
-// export/realty-groups/index.ts
-
-export type {
-  GroupCardLayout,
-  GroupMember,
-  GroupRating,
-  GroupReview,
-  GroupSidebarPlacement,
-  GroupsSortKey,
-  RealtyGroupCard,
-  SortDirection,
-} from "./types";
-
-export {
-  MOCK_AGENTS,
-  MOCK_LEADER,
-  MOCK_REVIEWS,
-  REALTY_GROUPS,
-} from "./mockData";
-
-export { cn, formatFounded, initialsOf } from "./lib/cn";
-
-export { Avatar } from "./components/Avatar";
-export { Badge } from "./components/Badge";
-export { Btn } from "./components/Btn";
-export { Checkbox } from "./components/Checkbox";
-export { DetailRow } from "./components/DetailRow";
-export { EmptyGroups } from "./components/EmptyGroups";
-export { FilterGroup } from "./components/FilterGroup";
-export { GroupCard } from "./components/GroupCard";
-export { GroupCover } from "./components/GroupCover";
-export { GroupLogo } from "./components/GroupLogo";
-export { MemberRow } from "./components/MemberRow";
-export { Pagination } from "./components/Pagination";
-export { StarPicker } from "./components/StarPicker";
-export { StarRating } from "./components/StarRating";
-
-export { GroupDetailPage } from "./pages/GroupDetailPage";
-export { GroupsPage } from "./pages/GroupsPage";
