+++ frontend/src/components/realty/browse/components/Badge.tsx
@@ -1,8 +1,7 @@
-// export/realty-groups/components/Badge.tsx
 "use client";
 
 import type { ReactNode } from "react";
-import { cn } from "../lib/cn";
+import { cn } from "@/lib/cn";
 
 type Tone =
   | "brand"
