+++ frontend/src/components/realty/browse/components/DetailRow.tsx
@@ -1,4 +1,3 @@
-// export/realty-groups/components/DetailRow.tsx
 "use client";
 
 import type { ReactNode } from "react";
