+++ frontend/src/components/realty/browse/components/GroupCover.tsx
@@ -1,4 +1,3 @@
-// export/realty-groups/components/GroupCover.tsx
 "use client";
 
 interface GroupCoverProps {
